Applications Refactor and demo ready
